Open files in DebatingActivity rather than FormatChooserActivity
