+++ app/src/main/AndroidManifest.xml
@@ -40,8 +40,15 @@
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+            <intent-filter android:label="@string/intentFilter_importStyle"
+                tools:ignore="GoogleAppIndexingWarning">
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.DEFAULT" />
+                <data android:mimeType="text/xml" />
+                <data android:mimeType="application/xml" />
+                <data android:mimeType="text/plain" android:pathPattern=".*\\.xml" />
             </intent-filter>
         </activity>
         <activity
@@ -59,14 +66,6 @@
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".DebatingActivity" />
-            <intent-filter android:label="@string/intentFilter_importStyle"
-                tools:ignore="GoogleAppIndexingWarning">
-                <action android:name="android.intent.action.VIEW" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="text/xml" />
-                <data android:mimeType="application/xml" />
-                <data android:mimeType="text/plain" android:pathPattern=".*\\.xml" />
-            </intent-filter>
         </activity>
         <activity
             android:name=".PrepTimeBellsEditActivity"
