+++ app/src/main/res/values/format_chooser.xml
@@ -34,9 +34,7 @@
     <string name="formatChooser_share_error_builtInFile">You can\'t share built-in formats.</string>
     <string name="formatChooser_share_error_notFound">File %1$s could not be found.</string>
     <string name="formatChooser_share_error_generic">There was an error sharing this file.</string>
-    <string name="formatChooser_view_error_generic">There was an error opening this file.</string>
     <string name="formatChooser_share_chooserTitle">Share debate format using…</string>
-    <string name="formatChooser_snackbar_importSuccessful">Imported file %1$s</string>
 
     <string name="viewFormat_regionLabel">Region</string>
     <string name="viewFormat_levelLabel">Level</string>
