+++ app/src/main/res/values/strings.xml
@@ -66,6 +66,10 @@
     <string name="mainScreen_snackbar_editTextDiscardChangesInfo_limitedBelow">To discard changes, press the back button. (You can\'t set the time to an overtime value.)</string>
     <string name="mainScreen_snackbar_editTextDiscardChangesInfo_limitedAbove">To discard changes, press the back button. (You can\'t set the time to greater than 23:59.)</string>
     <string name="mainScreen_snackbar_beamNonExternalFile">You can\'t use Android Beam to share built-in styles.</string>
+
+    <!-- Sharing debate formats -->
+    <string name="importDebateFormat_snackbar_error_generic">There was an error opening this file.</string>
+    <string name="importDebateFormat_snackbar_success">Imported file %1$s</string>
     <string name="mainScreen_snackbar_error_generic">There was an error beaming this debate format file.</string>
 
     <!-- Menu -->
